Add Data package and change file structure
