+++ app/Data/models/__init__.py
@@ -1,0 +1,6 @@
+def main():
+    pass
+
+
+if __name__ == "__main__":
+    main()
