+++ app/Data/models/contactpersons.py
@@ -1,6 +1,6 @@
 from sqlalchemy import Column, Integer, String
 from sqlalchemy.orm import relationship
-from db import Base
+from Data.db import Base
 
 
 class ContactPerson(Base):
