+++ app/Data/models/customers.py
@@ -1,6 +1,6 @@
 from sqlalchemy import Column, ForeignKey, Integer, String
 from sqlalchemy.orm import relationship
-from db import Base
+from Data.db import Base
 
 
 class Customer(Base):
