+++ app/Data/models/customertypes.py
@@ -1,16 +1,16 @@
 from sqlalchemy import Column, Integer, String
 from sqlalchemy.orm import relationship
-from db import Base
+from Data.db import Base
 
 
-class PartnerType(Base):
+class CustomerType(Base):
 
-    __tablename__ = "partner_types"
+    __tablename__ = 'customer_types'
 
     id = Column(Integer, primary_key=True)
     name = Column(String(45), nullable=False)
 
-    partners = relationship('Partner', back_populates='partner_type')
+    customers = relationship('Customer', back_populates='customer_type')
 
     def __repr__(self):
         pass
