+++ app/Data/models/matching_products.py
@@ -1,4 +1,4 @@
-from db import Base
+from Data.db import Base
 from sqlalchemy import Column, Integer, ForeignKey, Table
 
 matching_products = Table(
