+++ app/Data/models/partnertypes.py
@@ -1,16 +1,16 @@
 from sqlalchemy import Column, Integer, String
 from sqlalchemy.orm import relationship
-from db import Base
+from Data.db import Base
 
 
-class CustomerType(Base):
+class PartnerType(Base):
 
-    __tablename__ = 'customer_types'
+    __tablename__ = "partner_types"
 
     id = Column(Integer, primary_key=True)
     name = Column(String(45), nullable=False)
 
-    customers = relationship('Customer', back_populates='customer_type')
+    partners = relationship('Partner', back_populates='partner_type')
 
     def __repr__(self):
         pass
