+++ app/Data/models/products.py
@@ -1,4 +1,4 @@
-from db import Base
+from Data.db import Base
 from sqlalchemy import Column, Integer, String, ForeignKey, types
 from sqlalchemy.orm import relationship
