+++ app/main.py
@@ -1,17 +1,17 @@
-from db import Base, engine, session
-from models.products import Product
-from models.carmodels import CarModel
-# from models.contactpersons import ContactPerson
-# from models.customercar import CustomerCar
-# from models.customers import Customer
-# from models.customertypes import CustomerType
-# from models.employees import Employee
-# from models.orders import Order
-# from models.partnertypes import PartnerType
-# from models.partners import Partner
-# from models.stores import Store
-# from models.orderedproducts import OrderedProduct
-from models.matching_products import matching_products
+from Data.db import Base, engine
+from Data.models.products import Product
+from Data.models.carmodels import CarModel
+# from Data.models.contactpersons import ContactPerson
+# from Data.models.customercar import CustomerCar
+# from Data.models.customers import Customer
+# from Data.models.customertypes import CustomerType
+# from Data.models.employees import Employee
+# from Data.models.orders import Order
+# from Data.models.partnertypes import PartnerType
+# from Data.models.partners import Partner
+# from Data.models.stores import Store
+# from Data.models.orderedproducts import OrderedProduct
+from Data.models.matching_products import matching_products
 
 
 def main():
